Switch to code-fencing with language annotations
This isn't [Gruber's legacy Markdown](https://daringfireball.net/projects/markdown/syntax#precode) syntax, but is a widely-supported _de-facto_ standard for annotating blocks of code in a less ambiguous way. It is [defined in the CommonMark specification](http://spec.commonmark.org/0.27/#fenced-code-blocks) which many Markdown parsers are beginning/in-the-process-of adopting, and is a core feature of [GitHub-Flavored Markdown](https://github.com/github/cmark) (GFM).