Add item authors to spec
It is common for articles or podcasts to have multiple authors.
Examples include The New York Times and The New Yorker:

https://github.com/brentsimmons/JSONFeed/issues/6

Currently, the JSON feed spec allows one `author` object
on the top-level feed and one on each item:

```json
author: {
  name: "Alice",
  url: "https://example.org/alice",
  avatar: "https://example.org/alice.png",
}
```

To provide an item with multiple authors,
the author names could be concatenated and URLs and avatars removed:

```json
author: {
  name: "Alice and Bob",
}

Or, the author names could be concatenated
and combined URLs and avatars provided:

```json
author: {
  name: "Alice and Bob",
  url: "https://example.org/alice-and-bob",
  avatar: "https://example.org/alice-and-bob.png",
}
```

Those use cases might work well when the author is a team:

```json
author: {
  name: "Engineering Team",
  url: "https://example.org/team/engineering",
  avatar: "https://example.org/team/engineering.png",
}
```

This feels more awkward for use cases of blog posts / news articles
with multiple authors, or podcasts with multiple hosts where
the publisher wants to provide links for each author's individual
identity, such as to a listing of that author's other works.

In those cases, this structure feels nice to me:

```json
authors: [
  {
    name: "Alice",
    url: "https://example.org/alice",
  },
  {
    name: "Bob",
    url: "https://example.org/bob",
  }
]
```

I can imagine a natural-looking feed reader experience of:

```
Article Title
by Alice and Bob
May 6, 2018
```

This change adds `authors` to `items` and attempts to make it clear
to implementers how to handle the various present/missing states
for `author` and `authors`. Although all use cases can be covered by the
`authors` array (single author items can be modeled as a one item array),
it does not deprecate `author`. I leave that decision to project maintainers.

Also, this change does not add `authors` to the top-level feed.
In my experience, even on publications with multiple authors per item,
the top-level feed `author` has appropriately been the publishing
company or organization.